+++ public/index.html
@@ -206,7 +206,7 @@
         </div>
     </div>
     
-<script src="js/updateBook.js"></script>
+<script src="js/update-book.js"></script>
 </body>
 
 </html>
